FIX conda env list
