+++ benchopt/cli/completion.py
@@ -2,8 +2,8 @@ from pathlib import Path
 
 
 from benchopt.benchmark import Benchmark
-from benchopt.utils.misc import list_conda_envs
 from benchopt.utils.safe_import import skip_import
+from benchopt.utils.conda_env_cmd import list_conda_envs
 
 
 def get_benchmark(ctx, args, incomplete):
