+++ benchopt/utils/dependencies_mixin.py
@@ -1,9 +1,9 @@
 from ..config import RAISE_INSTALL_ERROR
 
 from .class_property import classproperty
-from .shell_cmd import install_in_conda_env
 from .shell_cmd import _run_shell_in_conda_env
-from .shell_cmd import shell_install_in_conda_env
+from .conda_env_cmd import install_in_conda_env
+from .conda_env_cmd import shell_install_in_conda_env
 
 
 class DependenciesMixin:
